% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twas2.R
\name{TWAS2}
\alias{TWAS2}
\title{Adaptive transcript mediated genome-wide association test using GWAS summary data}
\usage{
TWAS2(Z, R, W, rho = c((0:5/10)^2, 0.5, 1))
}
\arguments{
\item{Z}{summary Z-statistics for a set of variants}

\item{R}{the estimated variant LD matrix}

\item{W}{variant weight derived from outside reference panel of transcript and SNP data}

\item{rho}{sequence of weights assigned to the S-PrediXcan (summary statistics based PrediXcan) test}
}
\value{
\describe{
  \item{p.value}{ p-values for: PXA (adaptive test); PXV (VC test); PXU (S-PrediXcan) }
  \item{pvals}{ vector of p-values for all weighted tests }
  \item{rho.est}{ estimated optimal \eqn{\rho} value }
}
}
\description{
Take input of summary Z-statistics, variant LD matrix, transcript derived variant weight.
}
\examples{
library(IWAS)
## pairwise cor = 0.2
R = cor(matrix(rnorm(2e3),100,20)*sqrt(0.8)+rnorm(100)*sqrt(0.2))
Z = rnorm(20)*sqrt(0.8)+rnorm(1)*sqrt(0.2); W = rnorm(20)
TWAS2(Z,R,W)
Z[1:2] = Z[1:2] + 1*W[1:2]
TWAS2(Z,R,W)
}
\references{
Gamazon, E.R., Wheeler, H.E., Shah, K.P. et al. (2015) A gene-based association method for mapping traits using reference transcriptome data. Nat. Genet. 47, 1091–1098.

Gusev, A., Ko, A., Shi, H., et al. (2016) Integrative approaches for large-scale transcriptome-wide association studies. Nat. Genet. 48, 245–252. 

Wu,B., Guan,W., Pankow,J.S. (2016) On efficient and accurate calculation of significance p-values for sequence kernel association test of variant set. Annals of human genetics, 80(2), 123-135.

Wu,B., Guo,B. and Liu,N. (2017) A powerful and efficient statistical method for transcriptome-wide association test using GWAS summary data. tech rep.
}
